+++ src/app/components/credits/credits.scss
