+++ src/app/components/guindaille2-0/guindaille2-0.scss
