+++ src/app/components/map/search/search.scss
