+++ src/app/components/tuto/tuto.scss
